% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CppTrials-package.R
\docType{package}
\name{CppTrials-package}
\alias{CppTrials}
\alias{CppTrials-package}
\title{CppTrials: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
